# FatecAppBackend/FatecAppBackend.API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FatecApp API",
    "description": "Documentation of Api's endpoints",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "João Vitor",
      "url": "https://github.com/JoVi0li",
      "email": "jovioli.dev04@gmail.com"
    },
    "license": {
      "name": "Nothing",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/user/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create new User",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User props",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Domain.Commands.User.CreateUserCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
    "/api/v1/user/delete/{id}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete User",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
    "/api/v1/user/update": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update User",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User id",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Domain.Commands.User.UpdateUserCommand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Commands.GenericCommandsResult"
            }
          }
        }
      }
    },
...
    "/api/v1/user/id/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get User by id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Queries.GenericQueryResult"
            }
          }
        }
      }
    },
    "/api/v1/user/email/{email}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get User by email",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "email",
            "description": "User email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Queries.GenericQueryResult"
            }
          }
        }
      }
    },
    "/api/v1/user/name/{name}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get User by name",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "description": "User name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FatecAppBackend.Shared.Queries.GenericQueryResult"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "FatecAppBackend.Domain.Commands.User.CreateUserCommand": {
      "type": "object",
      "properties": {
        "notifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Flunt.Notifications.Notification"
          },
          "readOnly": true
        },
        "isValid": {
          "type": "boolean",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "photo": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "identityDocumentNumber": {
          "type": "string"
        },
        "identityDocumentPhoto": {
          "type": "string"
        },
        "gender": {
          "$ref": "#/definitions/FatecAppBackend.Shared.EnGender"
        },
        "validatedUser": {
          "type": "boolean"
        }
      }
    },
    "FatecAppBackend.Domain.Commands.User.UpdateUserCommand": {
      "type": "object",
      "properties": {
        "notifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Flunt.Notifications.Notification"
          },
          "readOnly": true
        },
        "isValid": {
          "type": "boolean",
          "readOnly": true
        },
        "updateUser": {
          "$ref": "#/definitions/FatecAppBackend.Shared.DTOs.User.UpdateUserDTO"
        }
      }
    },
...
    "FatecAppBackend.Shared.Commands.GenericCommandsResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": { }
      }
    },
...
    "FatecAppBackend.Shared.DTOs.User.UpdateUserDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "userCollegeId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "photo": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "identityDocumentNumber": {
          "type": "string"
        },
        "identityDocumentPhoto": {
          "type": "string"
        },
        "gender": {
          "$ref": "#/definitions/FatecAppBackend.Shared.EnGender"
        },
        "validatedUser": {
          "type": "boolean"
        }
      }
    },
...
    "FatecAppBackend.Shared.EnGender": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    },
...
    "FatecAppBackend.Shared.Queries.GenericQueryResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": { }
      }
    },
    "Flunt.Notifications.Notification": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}